alt42/orchestration/scripts/requirements.txt
# 문서-시스템 동기화 도구 의존성
# Python 3.10+

PyYAML>=6.0

